+++ src/DDTrace/Propagators/CurlHeadersMap.php
@@ -2,8 +2,11 @@
 
 namespace DDTrace\Propagators;
 
+use DDTrace\Configuration;
 use DDTrace\Propagator;
+use DDTrace\Sampling\PrioritySampling;
 use DDTrace\SpanContext;
+use DDTrace\Tracer;
 
 /**
  * A propagator that inject distributed tracing context in curl like indexed headers arrays:
@@ -11,6 +14,19 @@ use DDTrace\SpanContext;
  */
 final class CurlHeadersMap implements Propagator
 {
+    /**
+     * @var Tracer
+     */
+    private $tracer;
+
+    /**
+     * @param Tracer $tracer
+     */
+    public function __construct(Tracer $tracer)
+    {
+        $this->tracer = $tracer;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -29,6 +45,10 @@ final class CurlHeadersMap implements Propagator
                     === Propagator::DEFAULT_BAGGAGE_HEADER_PREFIX
             ) {
                 unset($carrier[$index]);
+            } elseif (substr($value, 0, strlen(Propagator::DEFAULT_SAMPLING_PRIORITY_HEADER))
+                === Propagator::DEFAULT_SAMPLING_PRIORITY_HEADER
+            ) {
+                unset($carrier[$index]);
             }
         }
 
@@ -37,6 +57,11 @@ final class CurlHeadersMap implements Propagator
 
         foreach ($spanContext as $key => $value) {
             $carrier[] = Propagator::DEFAULT_BAGGAGE_HEADER_PREFIX . $key . ': ' . $value;
+        }
+
+        $prioritySampling = $this->tracer->getPrioritySampling();
+        if (PrioritySampling::UNKNOWN !== $prioritySampling) {
+            $carrier[] = Propagator::DEFAULT_SAMPLING_PRIORITY_HEADER . ': ' . $prioritySampling;
         }
     }