+++ src/DDTrace/Sampling/AlwaysKeepSampler.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace DDTrace\Sampling;
+
+use DDTrace\Span;
+
+/**
+ * The simplest sampler, always providing a PrioritySampling::AUTO_KEEP option in any circumstance.
+ */
+class AlwaysKeepSampler implements Sampler
+{
+    /**
+     * @param Span $span
+     * @return int
+     */
+    public function getPrioritySampling(Span $span)
+    {
+        return PrioritySampling::AUTO_KEEP;
+    }
+}