+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file - [docs/chan
 - Tracing for the [legacy MongoDB extension](http://php.net/manual/en/book.mongo.php) for PHP 5 #166
 - Injecting distributed tracing headers in guzzle and curl requests #167
 - Possibility to autoload all integrations and to disable specific ones #168
+- Priority Sampling handling #169
 
 ### Fixed
 - "Undefined offset: 0" errors in ElasticSearch integration #165